% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Transforms the raw table file in something better.}
\usage{
build_table(arq)
}
\arguments{
\item{arq}{A string containing the path to the raw table.}
}
\value{
The raw table in tibble format with cleaned names and more intelligible tree format.
}
\description{
Transforms the raw table file in something better.
}
\examples{

library(tpur)
build_table(download_table("mov","estadual","1 grau"))
build_table(download_table("mov","federal","1 grau"))
}
