% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classes}
\alias{classes}
\title{Links do releases para as TPUs de classes}
\format{
Data frame:
\describe{
 \item{file}{Nome do arquivo que está no releases}
 \item{dt_inic}{Data de início do período de validade para a TPU}
 \item{dt_fim}{Data de fim do período de validade para a TPU}
 \item{release}{Link para a base .csv no releases}
 }
}
\usage{
classes
}
\description{
Uma base de resumo para os links do releases
}
\keyword{datasets}
