% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_table}
\alias{control_table}
\title{List of available tables and its links.}
\format{A data frame with 75 rows and 8 variables
\describe{
  \item{link}{URL for table download.}
  \item{table_type}{Type of table. Available values are "Assuntos","Classes","Movimentos"}
  \item{court}{Table's jurisdiction.}
  \item{level}{Table's jurisdiction level.}
  \item{status}{Status code of a GET requisition.}
}}
\source{
\url{https://www.cnj.jus.br/sgt/versoes.php}
}
\usage{
control_table
}
\description{
Dataset containing a list of the available tables, its links and
columns describing administrative boundaries.
}
\keyword{datasets}

