% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_classifier.R
\name{cria_bn}
\alias{cria_bn}
\title{Creates a bnlearn's graph based on a its edges.}
\usage{
cria_bn(edges)
}
\arguments{
\item{edges}{A 2 column matrix containg the edges of the graph.}
}
\value{
A bnlearn type graph.
}
\description{
Creates a bnlearn's graph based on a its edges.
}

