% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_page_on.R
\name{is_page_on}
\alias{is_page_on}
\title{Receives the status code of a GET requisition.}
\usage{
is_page_on(x)
}
\arguments{
\item{x}{The url to be requested.}
}
\value{
The status code of the get requisiton.
}
\description{
Receives the status code of a GET requisition.
}
\examples{
is_page_on("www.google.com")
}

