% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_classifier.R
\name{leaf_classifier}
\alias{leaf_classifier}
\title{Returns a leaf list from a Unified Processual Table.}
\usage{
leaf_classifier(pre_processed_table)
}
\arguments{
\item{pre_processed_table}{A Unified Processual Table processed with build_table.}
}
\value{
A list with the edges of the table.
}
\description{
Returns a leaf list from a Unified Processual Table.
}
