% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_fix.R
\name{level_fix}
\alias{level_fix}
\title{Replaces every NA in some columns by the last non null string observed.}
\usage{
level_fix(level)
}
\arguments{
\item{d}{a data frame.}
}
\value{
d with different columns names.
}
\description{
Replaces every NA in some columns by the last non null string observed.
}

