% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_fix.R
\name{name_fix}
\alias{name_fix}
\title{Replaces NA's names in a data frame columns list.}
\usage{
name_fix(d)
}
\arguments{
\item{d}{a data frame.}
}
\value{
d with different columns names.
}
\description{
Replaces NA's names in a data frame columns list.
}

