% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgt}
\alias{sgt}
\title{Links de todas as TPUs}
\format{
Data frame:
\describe{
 \item{id}{Código de identificação único da TPU}
 \item{atual}{Variável lógica se a TPU é atual ou não}
 \item{tipo}{Tipo de conteúdo da TPU (Assuntos, Classes ou Movimentos)}
 \item{tipo_abbr}{Abreviação do tipo de conteúdo da TPU (A, C ou M)}
 \item{data_versa}{Data da versão da TPU}
 \item{justica}{Justiça a que se refere a TPU}
 \item{tabela}{Ramificação da Justiça}
 \item{link}{Link da TPU}
 }
}
\usage{
sgt
}
\description{
Uma base de dados com todos os links de todas as TPUs, extraídos diretamente 
do site do Conselho Nacional de Justiça (CNJ) do Sistema de Gestão Processual (SGT).
}
\keyword{datasets}
