% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_links}
\alias{table_links}
\title{List of available tables and its links.}
\format{A data frame with 75 rows and 8 variables
\describe{
  \item{link}{URL for table download.}
  \item{table_type}{Classes, subjects of moves.}
  ...
}}
\source{
\url{https://www.cnj.jus.br/sgt/versoes.php}
}
\usage{
table_links
}
\description{
Dataset containing a list of the available tables, its links and
columns describing administrative boundaries.
}
\keyword{datasets}

