% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_tpu}
\alias{table_tpu}
\title{List of UPT's}
\format{A data frame with 48 rows and 10 variables:
\describe{
  \item{tpu}{Nested data_frames representing every download UPT.}
  \item{file_name}{Mnemonical identifier of the (table_type, court, level) triple.}
}}
\source{
\url{https://www.cnj.jus.br/sgt/versoes.php}
}
\usage{
table_tpu
}
\description{
Dataset containing every downloaded and pre-processed UPT
nested in a data_frame column.
}
\keyword{datasets}

