% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-assunto.R
\name{tpu_assunto_download}
\alias{tpu_assunto_download}
\title{Baixar html assunto TPU}
\usage{
tpu_assunto_download(sig, path = "data-raw/tpu/A")
}
\arguments{
\item{sig}{Sigla que se deseja baixar. A lista de siglas está na base sgt. Esta função recebe apenas as siglas iniciadas com "A"}

\item{path}{Diretório em que o html será salvo}
}
\value{
Retorna o path dos arquivos salvos
}
\description{
Baixa um arquivo html na pasta indicada
}
