% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-assunto.R
\name{tpu_assunto_parse}
\alias{tpu_assunto_parse}
\title{Parse html assunto TPU}
\usage{
tpu_assunto_parse(file)
}
\arguments{
\item{file}{Recebe o path de um arquivo html}
}
\value{
Retorna um data frame, com 16 colunas
}
\description{
Realiza o parse do html de um arquivo de assunto das TPUs
}
