% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{tpu_assunto_prob}
\alias{tpu_assunto_prob}
\title{Probability of leaf nodes}
\usage{
tpu_assunto_prob(d_litig, leafs, tpu, model = bn_indep())
}
\arguments{
\item{d_litig}{base de dados processual.}

\item{leafs}{nomes das folhas a serem incluidas na pesquisa.}

\item{tpu}{TPU a ser utilizada como base do modelo.}

\item{model}{modelo de classe. Defaults to independent model (MCAR).}
}
\description{
Probability of leaf nodes
}
