% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-assunto.R
\name{tpu_assunto_read}
\alias{tpu_assunto_read}
\title{Ler assuntos da TPU}
\usage{
tpu_assunto_read(busca = NULL, ini = Sys.Date(), fim = Sys.Date())
}
\arguments{
\item{busca}{Palavra a ser buscada nos assuntos}

\item{ini}{Data de início, por default o dia de hoje}

\item{fim}{Data de fim, por default o dia de hoje}
}
\value{
Retorna tibble com 5 colunas:
}
\description{
Retorna os códigos de assuntos das TPUs por período
}
\examples{
tpu_assunto_read("medicamento", ini = "2023-01-01", fim = "2023-02-01")

}
