% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-assunto.R
\name{tpu_assunto_tidy}
\alias{tpu_assunto_tidy}
\title{Arruma uma base de assuntos}
\usage{
tpu_assunto_tidy(da)
}
\arguments{
\item{da}{Recebe um data frame}
}
\value{
Retorna um data frame arrumado, com 16 colunas
}
\description{
Realiza a arrumação de uma base de dados contendo os assuntos de uma TPU
}
