% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-classe.R
\name{tpu_classe_download}
\alias{tpu_classe_download}
\title{Baixar html classe TPU}
\usage{
tpu_classe_download(sig, path = "data-raw/tpu/C")
}
\arguments{
\item{sig}{Sigla que se deseja baixar. A lista de siglas está na base sgt. Esta função recebe apenas as siglas iniciadas com "C"}

\item{path}{Diretório em que o html será salvo}
}
\value{
Retorna o path dos arquivos salvos
}
\description{
Baixa um arquivo html na pasta indicada
}
