% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-classe.R
\name{tpu_classe_parse}
\alias{tpu_classe_parse}
\title{Parse html classe TPU}
\usage{
tpu_classe_parse(file)
}
\arguments{
\item{file}{Recebe o path de um arquivo html}
}
\value{
Retorna um data frame, com 17 colunas
}
\description{
Realiza o parse do html de um arquivo de classe das TPUs
}
