% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-classe.R
\name{tpu_classe_read}
\alias{tpu_classe_read}
\title{Ler classes da TPU}
\usage{
tpu_classe_read(busca = NULL, ini = Sys.Date(), fim = Sys.Date())
}
\arguments{
\item{busca}{Palavra a ser buscada nas classes}

\item{ini}{Data de início, por default o dia de hoje}

\item{fim}{Data de fim, por default o dia de hoje}
}
\value{
Retorna tibble com 5 colunas:
}
\description{
Retorna os códigos de classes das TPUs por período
}
\examples{
tpu_classe_read("execução", ini = "2023-01-01", fim = "2023-02-01")

}
