% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu-classe.R
\name{tpu_classe_tidy}
\alias{tpu_classe_tidy}
\title{Arruma uma base de classes}
\usage{
tpu_classe_tidy(da)
}
\arguments{
\item{da}{Recebe um data frame}
}
\value{
Retorna um data frame arrumado, com 17 colunas
}
\description{
Realiza a arrumação de uma base de dados contendo as classes de uma TPU
}
