\name{tpur-package}
\alias{tpur-package}
\alias{tpur}
\docType{package}
\title{
\packageTitle{tpur}
}
\description{
\packageDescription{tpur}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tpur}
\packageIndices{tpur}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tpur}

Maintainer: \packageMaintainer{tpur}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
