% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpur.R
\docType{package}
\name{tpur}
\alias{tpur}
\alias{tpur-package}
\title{tpur: A package for acessing the Unified Processual Tables from the brazillian National Justice Council.}
\description{
tpur: A package for acessing the Unified Processual Tables from the brazillian National Justice Council.
}
\section{tpur functions}{

funcoes
}

