% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_fix.R
\name{type_fix}
\alias{type_fix}
\title{Transforms an expression in one of the available table_type's}
\usage{
type_fix(type)
}
\arguments{
\item{type}{The base expression.}
}
\value{
The closes available table_type
}
\description{
Transforms an expression in one of the available table_type's
}
