% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_fix.R
\name{court_fix}
\alias{court_fix}
\title{Transforms an expression in one of the available courts}
\usage{
court_fix(court)
}
\arguments{
\item{court}{The base expression.}
}
\value{
The closes available court.
}
\description{
Transforms an expression in one of the available courts
}
