% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{download_table}
\alias{download_table}
\title{Downloads the Unified Processual Table.}
\usage{
download_table(type, crt, lvl, arq = tempfile(), return_file = T)
}
\arguments{
\item{type}{String containing the type of the table.}

\item{crt}{String containing the court of the table.}

\item{lvl}{String containing the level of the table.}
}
\value{
The path of the download file.
}
\description{
Downloads the Unified Processual Table.
}
\examples{
download_table("mov","estadual","1º grau")
}

