% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_fix.R
\name{level_fix}
\alias{level_fix}
\title{Transforms an expression in one of the available jurisdiction levels.}
\usage{
level_fix(level)
}
\arguments{
\item{level}{The base expression.}
}
\value{
The closes available jurisdiction level.
}
\description{
Transforms an expression in one of the available jurisdiction levels.
}
