% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpur.R
\docType{package}
\name{tpur}
\alias{tpur}
\alias{tpur-package}
\title{tpur: Brazilian Unified Procedural Tables in R}
\description{
Download and access tables containing general purpose classifications for brazilian legal processes, providing some pre-processing as well
}
\section{tpur functions}{

funcoes
}

